Moved HttpResponseFallback to the agent
